Allow GPIO 9 and 10 for waveform generation (#5055)
* Allow GPIO 9 and 10 for waveform generation

While most ESP8266 modules use quad-io mode for their SPI flash ROM,
there are some which only use dual-io mode.  Allow the unused pins
(GPIO 9 and 10) to have waveforms generated on them.  Should the user try
this on a quad-io mode board, expect very bad things to happen.

* Add variant for 8285 to init GPIO 9/10

The 8285 only has 2-bit flash IO, so the other two pins can be used as
inputs (9/10).  Set them to input to mirror the way other pins are set
up.

* Update waveform gen to only allow 9/10 on 8285

Update the common.h in both generic (remove TODO, it's done!) and the
8285 variant to make isFlashInterfacePin macro correct.  Use that macro
to disable pins 9 and 10 in the common, non-8285 case.