Accepted WebServer Query Arguments not containing an equals sign
Fixes several issues and also Alexa queries.

Previous behavior was to completely ignore the argument (if not contains an equals sign), yet still include it in the argument count.
(The use of "key=value" arguments is a convention, not an actual requirement in URLs.)

For example, Amazon Echo sends a put-request with content-type "application/x-www-form-urlencoded" set in the headers but sends out json encoded data. _parseArguments fails to parse the arguments (which is correct) but completly discards the data. This change checks if the parser picked something up and if it didn't, it just adds it as plain data to the arguments.