Use Python JSON to format packages.json file
The packages JSON file which includes the boards, tools, etc. and needs to
have consistent formatting to be reproducible.  The current boards.txt.py
uses a REGEX to string-replace a bit of it, but that bit has a different
indent than the rest of the file.

Use Python's JSON writer to format the whole file repeatably.