Add Small String Optimization, min String size
Reduce String memory overhead from 24 bytes to 16 bytes by limiting the
maximum string length to <64Kbytes (which is larger than heap so no
effective problem).

Add Small String Optimization, SSO, which instead of allocating pointers
to small strings on the heap will store the string in place of the
pointer in the class.  This should reduce memory fragmentation as
Strings of 0-3 characters will not need extra malloc()s.

No user code changes should be required to work with this optimization.