Add host test, make SSO 8 bytes, remove magic num
Memory fragmentation is worse than not saving any RAM per String instance,
so extend the SSO size to 8 bytes (7 chars + '\0'), and making the total
String size the same size as it was before.

Remove any hardcoded SSO size and allow configuration via an enum, and
replace the magic max-capacity with an enum tor clarity.

Add a host test that verifies that no memory is allocated until a
full 8 characters are assigned to a string, as well as checking all
intermediate values.