Fix memory corruption on x64
Because pointers are 8 bytes (and 8-bytes aligned) on x64, the structure
used to store SSO needs to make sure the terminal \0 won't overwrite
some of that buffer.  Adjust the SSOSIZE dynamically depending on the
size of the struct ptr.

Refactor to add a setBuffer method, like the setConfig, setSSO, etc.

Fix some issues found when using SSO strings and functions which modify
len() dynamically.