Fix stack smashing error on 64b, undo setLen 0
When pointers are 8 bytes long, the size of a String is larger than 16
chars.  Increase the allocated array we're using in the test to avoid a
"stack smashing" error.

Undo setLen actually writing a 0 in memory when called in SSO mode. The
semantics of setLen are such that it is called as part of a multi-step
process to update a String, and when it is called it does NOT normally
write a 0 to the buffer at the length indicated.  Doing so in SSO mode
and 64b OSes exposed the error, now cleared up.