Fix PIO toolchain by ensuring IRQ vectors in link
PIO uses GCC in a way that causes the compiler to optimize away the IRQ
vectors (since they're not used in our code itself) in the final build
stage.

We can avoid this by forcing all of libmain to be included in the final
object.  This is a small file, and almost completely related to IRQs, so
space impact is minimal.

Fixes #6087 and
https://github.com/platformio/platform-espressif8266/issues/147