Support PIO Wl-T and Arduino -T linking properly
The interrupt vectors in IRAM are ommitted when there is a PROVIDE
statement in the linker control files when using the PIO method of
-Wl,-T<linkfile>.

Drop the PROVIDES (they're in RAM anyway and not ROM related), and
add the required "-u"s to the P{IO build script.

Should have no iumpact on the Arduino side.

Fixes #6087