Add mklittlefs and filename tags to releases
* Add timestamp to filenames to fix json issues

When GITs of the same tag are rebuilt, the binaries are different, but
the filenames would be the same so the Arduino tools would think they
already had the file downloaded.  But when they checked the signature of
the (old) downloaded file, it wouldn't match that (of the new one) in the
JSON.

Append a timestamp to the filenames, in seconds since the epoch, so that
different builds of the same GIT will have different names.

* Add mklittlefs binary to toolchain release, update JSON